refactor: 컨트롤러의 비즈니스 로직·DB 접근 책임 분리
