--- src/main/java/roomescape/ReservationTimeController.java
@@ -1,55 +1,0 @@
-package roomescape;
-
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import org.springframework.http.ResponseEntity;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-
-@Controller
-public class ReservationTimeController {
-    private final JdbcTemplate jdbcTemplate;
-
-    public ReservationTimeController(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
-
-    @GetMapping("/times")
-    public ResponseEntity<List<ReservationTime>> read() {
-        List<ReservationTime> times = jdbcTemplate.query(
-                "SELECT id, start_at FROM reservation_time",
-                (rs, rowNum) -> new ReservationTime(
-                        rs.getLong("id"),
-                        rs.getString("start_at")
-                )
-        );
-        return ResponseEntity.ok(times);
-    }
-
-    @PostMapping("/times")
-    public ResponseEntity<ReservationTime> create(@RequestBody ReservationTime reservationTime) {
-        SimpleJdbcInsert insert = new SimpleJdbcInsert(jdbcTemplate)
-                .withTableName("reservation_time")
-                .usingGeneratedKeyColumns("id");
-
-        Map<String, Object> params = new HashMap<>();
-        params.put("start_at", reservationTime.getStartAt());
-
-        Long id = insert.executeAndReturnKey(params).longValue();
-
-        return ResponseEntity.ok(new ReservationTime(id, reservationTime.getStartAt()));
-    }
-
-    @DeleteMapping("/times/{id}")
-    public ResponseEntity<ReservationTime> delete(@PathVariable Long id) {
-        jdbcTemplate.update("DELETE FROM reservation WHERE id = ?", id);
-        return ResponseEntity.ok().build();
-    }
-}
