+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -1,0 +1,37 @@
+package roomescape.controller;
+
+import java.util.List;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import roomescape.domain.ReservationTime;
+import roomescape.service.ReservationTimeService;
+
+@Controller
+public class ReservationTimeController {
+    private final ReservationTimeService reservationTimeService;
+
+    public ReservationTimeController(ReservationTimeService reservationTimeService) {
+        this.reservationTimeService = reservationTimeService;
+    }
+
+    @GetMapping("/times")
+    public ResponseEntity<List<ReservationTime>> read() {
+        return ResponseEntity.ok(reservationTimeService.findAll());
+    }
+
+    @PostMapping("/times")
+    public ResponseEntity<ReservationTime> create(@RequestBody ReservationTime reservationTime) {
+        return ResponseEntity.ok(reservationTimeService.save(reservationTime));
+    }
+
+    @DeleteMapping("/times/{id}")
+    public ResponseEntity<Void> delete(@PathVariable Long id) {
+        reservationTimeService.delete(id);
+        return ResponseEntity.ok().build();
+    }
+}
